pyproject.toml
[tool.poetry]
name = "channels-nats"
version = "0.0.1"
description = "NATS interface server for django-channels"
authors = ["Eirik Martiniussen Sylliaas <eirik@sylliaas.no>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
asgiref = "^3.2.5"
asyncio-nats-client = "^0.10.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.7.9"
mypy = "^0.770"

[tool.poetry.scripts]
channels-nats = 'channels_nats.cli:CommandLineInterface.entrypoint'
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

